refactor(promql): split the fused fold by DataFusion operator
fold.rs held the range expression, the aggregate, the projection, and the
partition collector; each now sits where a DataFusion reader expects it:
range_expr.rs (PhysicalExpr), aggregate.rs (partial then final),
project.rs (one output per series; map_sources becomes project), and
collect_partitioned in the module root. Moves only.
